==> AverageEmbeddingMethod/2_CreateArticleEmbedding.ipynb <==
Creates article embeddings for all articles

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

adf = pd.read_parquet('./data/articles.parquet')

adf.head(1)

adf['detail_desc'].fillna('no description given', inplace=True)

adf['perceived_colour_value_name'].value_counts().head(20)

adf['garment_group_name'].value_counts().head(20)

adf.head()

adf['prod_desc_custom'] = adf['detail_desc'] + '. It has a ' + adf['graphical_appearance_name'] + ' print on it.' + ' It has a ' + \
                        adf['perceived_colour_value_name'] + ' ' +  adf['colour_group_name'] + ' colour.'

# adf['prod_desc_custom'] = adf['product_type_name'] + ' named as ' + adf['prod_name'] + '. It has a ' + adf['graphical_appearance_name'] + ' print on it.' + ' It has a ' + \
#                         adf['perceived_colour_value_name'] + ' ' +  adf['colour_group_name'] + ' colour'\
#                         '. It is from the ' + adf['index_name'] + ' group and ' + adf['garment_group_name'] + ', ' + \
#                         adf['section_name'] + ' section. Description: ' + adf['detail_desc']

adf['prod_desc_custom'].iloc[np.random.randint(low=0, high=len(adf) - 1)]

adf['prod_desc_custom'].isna().sum()

model = hub.load("https://tfhub.dev/google/universal-sentence-encoder/4")

# adf['prod_desc_custom'] = adf['prod_desc_custom'].astype(str)

xx = model(['testing query'])

texts = adf['prod_desc_custom'].values.tolist()

# calculates end index for a particular iteration for looping through documents in batches
def calcEndIdx(start_idx, batch_size, ndocs):
    end_idx = start_idx + batch_size
    end_idx = ndocs if end_idx > ndocs - 1 else end_idx
    return end_idx

embeddings = [] # empty array to store embeddings as we iterate through docs
BATCH_SIZE = 64
NUM_DOCS = len(texts)

for start_idx in tqdm(range(0, NUM_DOCS, BATCH_SIZE)):
    end_idx = calcEndIdx(start_idx, BATCH_SIZE, NUM_DOCS)
    curr_embeddings = model(texts[start_idx:end_idx]).numpy()
    embeddings.append(curr_embeddings)
    
embeddings = np.concatenate(embeddings) # convert batched arrays to shape (N, Vector Size)

len(texts)

aids = adf['article_id'].values

emb_map = {}
for i in range(len(texts)):
    emb = embeddings[i]
    emb_map[aids[i]] = emb

len(emb_map)

len(aids)

with open('./data/emb_map.pkl', 'wb') as handle:
    pkl.dump(emb_map, handle, protocol=3)
