==> AverageEmbeddingMethod/3_CreateTargetEmbedding.ipynb <==
### Create Target Embedding for customers who bought

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

with open('./data/emb_map.pkl', 'rb') as handle:
    emb_map = pkl.load(handle)

adf = pd.read_parquet('./data/articles.parquet')

tdf = pd.read_parquet('./data/train.parquet')

bdf = pd.read_parquet('./data/bought_articles_in_order.parquet')

bdf_dict = bdf.set_index('cust_id').to_dict(orient='dict')['article_id']

bdf.head()

cids = bdf['cust_id'].unique()

ctarget_map = {}

for cid in tqdm(cids, total=len(cids)):
    keys = bdf_dict[cid][:12]
    embs = [emb_map[x] for x in keys]
    embs = np.array(embs)
    cemb = embs.mean(axis=0)
    cemb /= np.linalg.norm(cemb)
    ctarget_map[cid] = cemb

bdf_dict[13]

f = [693242018, 661794006, 763037004, 640176008, 688326010]

af = adf[adf['article_id'].isin(f)].copy()

akeys = list(emb_map.keys())
aembs = list(emb_map.values())

aembs = np.array(aembs)

akeys = np.array(akeys)

def calculateSimilarity(queryVector):
    scores = queryVector.dot(aembsT).flatten() # cosine similarity scores. (assumes vectors are normalized to unit length)
    return scores

aembsT = aembs.T

def rankPostsBySimilarity(posts_ids, scores):
    return sorted([(post_id, score) for (post_id, score) in zip(posts_ids, scores)], key=lambda x:x[1], reverse=True)[:12]

p = ctarget_map[13]

sims = calculateSimilarity(p)
temp = np.argpartition(sims, -12)[-12:]
r = rankPostsBySimilarity(akeys[temp], sims[temp])

r

adf[adf['article_id'] == 515262001]

af

with open('./data/ctarget_map.pkl', 'wb') as handle:
    pkl.dump(ctarget_map, handle, protocol=3)
