==> AverageEmbeddingMethod/5_EmbModel.ipynb <==
### Trains a neural network to predict a 512 dimensional vector

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

from tensorflow import keras
from tensorflow.keras import layers
import tensorflow as tf
import tensorflow_addons as tfa
from tensorflow.keras import regularizers

with open('./data/cemb_map.pkl', 'rb') as handle:
    cemb_map = pkl.load(handle)
    
with open('./data/ctarget_map.pkl', 'rb') as handle:
    ctarget_map = pkl.load(handle)

len(cemb_map)

len(ctarget_map)

keys = list(cemb_map.keys())

for k in keys:
    if k not in ctarget_map:
        print(k)

X, Y = [], []

for key in tqdm(keys, total=len(keys)):
    X.append(cemb_map[key])
    Y.append(ctarget_map[key])

X = np.array(X)
Y = np.array(Y)

X.shape

Y.shape

from sklearn.model_selection import train_test_split

X_train, X_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, random_state=42)

from sklearn.preprocessing import StandardScaler

scaler = StandardScaler()
scaler.fit(X_train)

X_train = scaler.transform(X_train)
X_test = scaler.transform(X_test)

with open('./data/scaler.pkl', 'wb') as handle:
    pkl.dump(scaler, handle, protocol=3)

BATCH_SIZE = 1

fold_train_loss = []
fold_val_loss = []

train_dataset = tf.data.Dataset.from_tensor_slices((X_train, y_train))
val_dataset = tf.data.Dataset.from_tensor_slices((X_test, y_test))

train_dataset = train_dataset.batch(BATCH_SIZE)
val_dataset = val_dataset.batch(64)

train_dataset.prefetch(tf.data.AUTOTUNE)
val_dataset.prefetch(tf.data.AUTOTUNE)

def build_and_compile_model():
    model = keras.Sequential([
        tf.keras.layers.InputLayer(
        input_shape=(X_train.shape[1],)),
        layers.Dense(25, activation='relu'),
        layers.Dense(12, activation='relu'),
        layers.Dense(12, activation='relu'),
        layers.Dense(12, activation='relu'),
        layers.Dense(512, activation='linear')
    ])

    model.compile(loss='mean_squared_error',
                optimizer=tf.keras.optimizers.SGD(0.01))
    return model

keras.backend.clear_session()
dnn = build_and_compile_model()
dnn.summary()

history = dnn.fit(
    train_dataset,
    validation_data=val_dataset,
    verbose=1, epochs=2, workers=1, shuffle=True)

dnn.save('./models/model1')

idx = 3

pred = dnn.predict(X_test[idx].reshape(1,-1))

L = tf.keras.losses.CosineSimilarity(axis=-1)

L(y_test[idx].reshape(1,-1), pred)

y_test[idx][:10]

pred[0][:10]
