==> AverageEmbeddingMethod/7_CreatePredEmbs.ipynb <==
### Generates the predictions to be submitted to kaggle

1. Generate embeddings for all users from training data
2. Use model to make predictions for users who have purchases >0
3. For users that dont have purchases, predict most popular from last week

import numba

import pandas as pd, os, numpy as np
import plotly.express as px
pd.options.display.max_columns = 50
import swifter, datetime, pickle as pkl
import tensorflow_hub as hub
from tqdm.notebook import tqdm

from tensorflow import keras
from tensorflow.keras import layers
import tensorflow as tf
import tensorflow_addons as tfa
from tensorflow.keras import regularizers

with open('./data/all_cemb_map.pkl', 'rb') as handle:
    all_cemb_map = pkl.load(handle)

with open('./data/scaler.pkl', 'rb') as handle:
    scaler = pkl.load(handle)

keys = list(all_cemb_map.keys())

values = list(all_cemb_map.values())

model = keras.models.load_model('./models/model1')

values_fs = np.array(values)

values_fs.shape

values_fs = scaler.transform(values_fs)

preds = model.predict(values_fs, verbose=1)

preds.shape

df = pd.read_parquet('./data/train.parquet')

d_end = datetime.datetime(2020, 9, 22).date()
d_start = datetime.datetime(2020, 9, 1).date()

tdf = df[(df['date'] >= d_start) & (df['date'] <= d_end)].reset_index(drop=True).copy()

uaids = tdf['article_id'].unique()

with open('./data/emb_map.pkl', 'rb') as handle:
    emb_map = pkl.load(handle)

f_emb_map = {}
for aid in uaids:
    f_emb_map[aid] = emb_map[aid]

len(f_emb_map)

len(uaids)

akeys = list(f_emb_map.keys())
aembs = list(f_emb_map.values())

aembs = np.array(aembs)

akeys = np.array(akeys)

aembsT = aembs.T

# def calculateSimilarity(queryVector):
#     scores = queryVector.dot(aembsT).flatten() # cosine similarity scores. (assumes vectors are normalized to unit length)
#     return scores

def calculateSimilarity2(queryVector):
    scores = np.linalg.norm(aembs - queryVector, axis=1).flatten()
    return scores

aembs.shape

def rankPostsBySimilarity(posts_ids, scores):
    return sorted([(post_id, score) for (post_id, score) in zip(posts_ids, scores)], key=lambda x:x[1], reverse=False)[:12]

import time

preds[0][:10]

preds[1][:10]

cpreds = []
start=time.time()
for p in tqdm(preds[:10], total=len(preds)):
    sims = calculateSimilarity2(p)
    temp = np.argpartition(sims, -12)[-12:]
    r = rankPostsBySimilarity(akeys[temp], sims[temp])
    rank_ids = [v[0] for v in r]
    
    preds_sep_f = ''
    for i in range(len(rank_ids)):
        val = '0' + str(rank_ids[i])
        if i != len(rank_ids) - 1:
            val += ' '
        preds_sep_f += val
    cpreds.append(preds_sep_f)
end = time.time()
print("total time taken: {}".format(end-start))
    # rank_ids = [v[0] for v in r]
    
    # preds_sep_f = ''
    # for i in range(len(rank_ids)):
    #     val = '0' + str(rank_ids[i])
    #     if i != len(rank_ids) - 1:
    #         val += ' '
    #     preds_sep_f += val
    # cpreds.append(preds_sep_f)

adf = pd.read_parquet('./data/articles.parquet')

adf[adf['article_id'] == 809590003]

pdf = pd.DataFrame()
pdf['cust_id'] = keys
pdf['prediction'] = cpreds

pdf.to_parquet('./data/cpreds.parquet')

pdf
